+++ js/dati.js
@@ -16,7 +16,7 @@
    A `twitch.domini` la generazione aggiunge sempre localhost e 127.0.0.1;
    il player ci mette da sé location.hostname.
 
-   Generato il 2026-09-20T12:33:34.986Z.
+   Generato il 2026-09-20T16:04:08.258Z.
    ===================================================================== */
 window.DATI = {
   "twitch": {